fix(NanoSpec): Add static_cast for double to long conversions (C4244) in OverlayDialog
- Fixed 2 C4244 warnings where double values were implicitly converted to long
- Added static_cast<long>() for coordinate assignments
- Ensures type safety and eliminates possible data loss warnings

+++ SRC/NanoSpec/OverlayDialog.cpp
@@ -3962,8 +3962,8 @@ CString strMsg;
 	STAGE_COORD coordLeftTop;
 	STAGE_COORD coordRightBottom;
 
-	coordLeftTop.lX = m_StageConfig.StageMaxCoord.dLeftX;
-	coordLeftTop.lY = m_StageConfig.StageMaxCoord.dUpY;
+	coordLeftTop.lX = static_cast<long>(m_StageConfig.StageMaxCoord.dLeftX);
+	coordLeftTop.lY = static_cast<long>(m_StageConfig.StageMaxCoord.dUpY);
 	coordRightBottom.lX = m_StageConfig.StageMaxCoord.dRightX;
 	coordRightBottom.lY = m_StageConfig.StageMaxCoord.dDownY;
